Fix last hanging test.
